fix nbid,cu_du_id to uint32 and mcc,mn to uint16

+++ examples/xApp/go/FlexPolicy/callbacks/cb_max_num_ues.go
@@ -33,7 +33,7 @@ func CallbackMaxNumUes(PolicyConfiguration policy.Configuration) {
 		currRanType := xapp.Get_e2ap_ngran_name(sm.E2Nodes.Get(i).GetId().GetXtype())
 
 		// - c) NbId
-		currNbId := int16(sm.E2Nodes.Get(i).GetId().GetNb_id().GetNb_id())
+		currNbId := uint32(sm.E2Nodes.Get(i).GetId().GetNb_id().GetNb_id())
 
 		// - d) CuDuId    TODO: maybe swig needs to be extended to support this
 		// currCuDuId := sm.E2Nodes.Get(i).GetId().GetCu_du_id().GetCu_du_id()
@@ -51,8 +51,8 @@ func CallbackMaxNumUes(PolicyConfiguration policy.Configuration) {
 
 		// STEP 1: Check if there is an idle slice and if no create it
 		e2nodeId := slice.E2NodeId{
-			Mcc:     int16(currMcc),
-			Mnc:     int16(currMnc),
+			Mcc:     currMcc,
+			Mnc:     currMnc,
 			NbId:    currNbId,
 			CuDuId:  0,
 			RanType: currRanType,
