+++ examples/xApp/go/FlexPolicy/non_rt_emu/max_num_ues/non_rt_max_num_ues_main.go
@@ -51,8 +51,8 @@ func main() {
 
 	rand.Seed(time.Now().UnixNano())
 
-	mcc := uint16(sm.Feedback.SliceFeedback[0].Mcc)
-	mnc := uint16(sm.Feedback.SliceFeedback[0].Mnc)
+	mcc := sm.Feedback.SliceFeedback[0].Mcc
+	mnc := sm.Feedback.SliceFeedback[0].Mnc
 	nbId := sm.Feedback.SliceFeedback[0].NbId
 	ranType := sm.Feedback.SliceFeedback[0].RanType
